requirements.txt
dash
flask-login
sqlalchemy
flask-sqlalchemy
dash_core_components

dash-bootstrap-components
pymysql
pandas
requests
python-dotenv
dash_daq
chardet
Werkzeug==2.2.3

reportlab
matplotlib
RotatingTextFile
yagmail
bcrypt
